requirements.txt
streamlit
groq
sentence-transformers
faiss-cpu
fastapi
uvicorn
gunicorn
requests
numpy
fastapi
uvicorn
python-multipart
faiss-cpu
numpy
sentence-transformers
pydantic
requests
gTTS
tqdm
google-genai
groq
protobuf<=3.20.3
aiofiles
transformers

